requirements.txt
click==6.7
ecdsa==0.13
Flask==0.12.2
grpcio==1.7.3
grpcio-tools==1.7.3
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
protobuf==3.5.0.post1
six==1.11.0
Werkzeug==0.13
